pyproject.toml
[project]
name = "joann"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "pytd>=1.9.0",
    "python-dotenv>=1.1.1",
]
